requirements.txt
asn1crypto==1.2.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
configparser==3.7.4
cryptography==2.8
dnspython==1.16.0
future==0.18.2
idna==2.8
numpy==1.16.4
oauthlib==3.1.0
pycparser==2.19
pymongo==3.7.2
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dotenv==0.10.3
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.3.0
schedule==0.6.0
six==1.13.0
urllib3==1.25.3
